RavenDB-18226 - create single processor for each EP
