+++ src/Raven.Server/Web/Studio/Sharding/ShardedStudioTasksHandler.cs
@@ -1,6 +1,8 @@
 ﻿using System.Threading.Tasks;
 using Raven.Server.Documents.Sharding.Handlers;
 using Raven.Server.Routing;
+using Raven.Server.ServerWide.Context;
+using Raven.Server.Web.Studio.Processors;
 
 namespace Raven.Server.Web.Studio.Sharding
 {
@@ -9,7 +11,8 @@ namespace Raven.Server.Web.Studio.Sharding
         [RavenShardedAction("/studio-tasks/periodic-backup/test-credentials", "POST")]
         public async Task TestPeriodicBackupCredentials()
         {
-            await StudioTasksHandler.TestPeriodicBackupCredentials(this);
+            using (var processor = new StudioTasksHandlerProcessorForTestPeriodicBackupCredentials<ShardedDatabaseRequestHandler, TransactionOperationContext>(this, ContextPool))
+                await processor.ExecuteAsync();
         }
     }
 }
